#2203 Adjusted documentation. Refactored scatter plot master behavior. Removed redundant behavior.
