+++ platform-web-ui/src/main/java/ua/com/fielden/platform/web/view/master/scatterplot/api/IScatterPlotAction.java
@@ -6,6 +6,8 @@ import ua.com.fielden.platform.web.centre.api.actions.EntityActionConfig;
 /**
  * Contract for assigning action to scatter plot on double-click event
  * @param <T>
+ *
+ * @author TG Team
  */
 public interface IScatterPlotAction<T extends AbstractEntity<?>> extends IScatterPlotDone<T>{
