+++ platform-web-ui/src/main/java/ua/com/fielden/platform/web/view/master/scatterplot/api/IScatterPlotEntityType.java
@@ -6,8 +6,16 @@ import ua.com.fielden.platform.entity.AbstractEntity;
  * Contract to specify the type of entities displayed on scatter plot.
  *
  * @param <T>
+ *
+ * @author TG Team
  */
 public interface IScatterPlotEntityType<T extends AbstractEntity<?>> {
 
+    /**
+     * Specifies the type of entities displayed on scatter plot.
+     *
+     * @param entityType
+     * @return
+     */
     IScatterPlotCategoryProperty<T> setChartEntityType (Class<? extends AbstractEntity<?>> entityType);
 }
