+++ platform-web-ui/src/main/java/ua/com/fielden/platform/web/view/master/scatterplot/api/IScatterPlotMasterBuilder.java
@@ -6,11 +6,13 @@ import ua.com.fielden.platform.entity.AbstractEntity;
  * A contract to configure scatter plot/
  *
  * @param <T>
+ *
+ * @author TG Team
  */
 public interface IScatterPlotMasterBuilder<T extends AbstractEntity<?>> {
 
     /**
-     * Start scatter plot configuration for specified entity type
+     * Start scatter plot configuration for specified master entity type
      *
      * @param entityType
      * @return
@@ -18,7 +20,7 @@ public interface IScatterPlotMasterBuilder<T extends AbstractEntity<?>> {
     IScatterPlotEntityType<T> forEntity(Class<T> entityType);
 
     /**
-     * Start scatter plot configuration for specified entity type. Also, this scatter plot master get saved when opened.
+     * Start scatter plot configuration for specified master entity type. Also, this scatter plot master get saved when opened.
      *
      * @param entityType
      * @return
