+++ platform-web-ui/src/main/java/ua/com/fielden/platform/web/view/master/scatterplot/api/IScatterPlotTooltip.java
@@ -4,9 +4,11 @@ import ua.com.fielden.platform.entity.AbstractEntity;
 import ua.com.fielden.platform.processors.metamodel.IConvertableToPath;
 
 /**
- * A contract for configuring tooltip on scatter plot
+ * A contract for configuring tooltip for scatter plot dots
  *
  * @param <T>
+ *
+ * @author TG Team
  */
 public interface IScatterPlotTooltip<T extends AbstractEntity<?>> extends IScatterPlotLegend<T> {
