+++ platform-web-ui/src/main/java/ua/com/fielden/platform/web/view/master/scatterplot/api/IScatterPlotValueProperty.java
@@ -4,14 +4,16 @@ import ua.com.fielden.platform.entity.AbstractEntity;
 import ua.com.fielden.platform.processors.metamodel.IConvertableToPath;
 
 /**
- * Contract to specify property name and axis range configuration that should be used to configure continuous axis of scatter plot (usually x-axis).
+ * Contract to specify property name and axis range configuration that should be used to configure continuous axis of scatter plot (x-axis).
  *
  * @param <T>
+ *
+ * @author TG Team
  */
 public interface IScatterPlotValueProperty<T extends AbstractEntity<?>> {
 
     /**
-     * Assigns property name (x-axis) and configuration of it's range
+     * Assigns property name and configuration of x-axis range
      *
      * @param propertyName
      * @return
