+++ platform-web-ui/src/main/java/ua/com/fielden/platform/web/view/master/scatterplot/api/IScatterPlotXAxisTitle.java
@@ -6,6 +6,8 @@ import ua.com.fielden.platform.entity.AbstractEntity;
  * A contract to specify x-axis title of scatter plot.
  *
  * @param <T>
+ *
+ * @author TG Team
  */
 public interface IScatterPlotXAxisTitle<T extends AbstractEntity<?>> extends IScatterPlotYAxisTitle<T> {
