+++ platform-web-ui/src/main/java/ua/com/fielden/platform/web/view/master/scatterplot/api/IScatterPlotYAxisTitle.java
@@ -6,6 +6,8 @@ import ua.com.fielden.platform.entity.AbstractEntity;
  * A contract to specify y-axis title of scatter plot.
  *
  * @param <T>
+ *
+ * @author TG Team
  */
 public interface IScatterPlotYAxisTitle<T extends AbstractEntity<?>> extends IScatterPlotTopMargin<T> {
