+++ platform-web-ui/src/main/java/ua/com/fielden/platform/web/view/master/scatterplot/api/implementation/ScatterPlotAxisRange.java
@@ -3,6 +3,11 @@ package ua.com.fielden.platform.web.view.master.scatterplot.api.implementation;
 import ua.com.fielden.platform.web.view.master.scatterplot.api.IScatterPlotAxisRangeProp;
 import ua.com.fielden.platform.web.view.master.scatterplot.api.IScatterPlotRangeConfig;
 
+/**
+ * Implements the scatter plot axis configuration (it's data source and property name in data source that is used to configure axis range)
+ *
+ * @author TG Team
+ */
 public class ScatterPlotAxisRange implements IScatterPlotRangeConfig, IScatterPlotAxisRangeProp {
 
     private String source = "";
@@ -15,7 +20,7 @@ public class ScatterPlotAxisRange implements IScatterPlotRangeConfig, IScatterPl
         return new ScatterPlotAxisRange("masterEntity");
     }
 
-    public ScatterPlotAxisRange (String source) {
+    private ScatterPlotAxisRange (String source) {
         this.source = source;
     }
