+++ platform-web-ui/src/main/java/ua/com/fielden/platform/web/view/master/scatterplot/api/implementation/ScatterPlotMaster.java
@@ -2,14 +2,11 @@ package ua.com.fielden.platform.web.view.master.scatterplot.api.implementation;
 
 import org.apache.commons.lang3.StringUtils;
 import ua.com.fielden.platform.basic.IValueMatcherWithContext;
-import ua.com.fielden.platform.dom.DomContainer;
 import ua.com.fielden.platform.dom.DomElement;
 import ua.com.fielden.platform.dom.InnerTextElement;
 import ua.com.fielden.platform.entity.AbstractEntity;
 import ua.com.fielden.platform.processors.metamodel.IConvertableToPath;
-import ua.com.fielden.platform.reflection.AnnotationReflector;
 import ua.com.fielden.platform.reflection.PropertyTypeDeterminator;
-import ua.com.fielden.platform.reflection.TitlesDescsGetter;
 import ua.com.fielden.platform.types.Money;
 import ua.com.fielden.platform.utils.EntityUtils;
 import ua.com.fielden.platform.utils.Pair;
@@ -19,14 +16,12 @@ import ua.com.fielden.platform.web.centre.api.resultset.impl.FunctionalActionEle
 import ua.com.fielden.platform.web.centre.api.resultset.impl.FunctionalActionKind;
 import ua.com.fielden.platform.web.interfaces.IRenderable;
 import ua.com.fielden.platform.web.view.master.api.IMaster;
-import ua.com.fielden.platform.web.view.master.chart.decker.api.impl.ChartLine;
 
 import java.util.*;
-import java.util.function.Function;
 import java.util.stream.Collectors;
 
 import static java.lang.String.format;
-import static java.util.Optional.ofNullable;
+import static java.util.Optional.*;
 import static org.apache.commons.lang3.StringUtils.isEmpty;
 import static org.apache.commons.lang3.StringUtils.join;
 import static ua.com.fielden.platform.reflection.TitlesDescsGetter.*;
@@ -38,29 +33,40 @@ import static ua.com.fielden.platform.web.view.master.EntityMaster.ENTITY_TYPE;
 import static ua.com.fielden.platform.web.view.master.EntityMaster.flattenedNameOf;
 import static ua.com.fielden.platform.web.view.master.api.impl.SimpleMasterBuilder.createImports;
 
+/**
+ * Represents the entity master for scatter plot.
+ *
+ * @param <T>
+ *
+ * @author TG Team
+ */
 public class ScatterPlotMaster<T extends AbstractEntity<?>> implements IMaster<T> {
 
     private final IRenderable renderable;
-    private final List<EntityActionConfig> actions = new ArrayList<>();
 
+    private final EntityActionConfig action;
+    /**
+     * Create new scatter plot according to given configuration
+     *
+     * @param scatterPlotMasterBuilder
+     */
     public ScatterPlotMaster(final ScatterPlotMasterBuilder<T> scatterPlotMasterBuilder) {
+        this.action = scatterPlotMasterBuilder.getAction();
 
         final LinkedHashSet<String> importPaths = new LinkedHashSet<>();
-
-        actions.add(scatterPlotMasterBuilder.getAction());
-        final Pair<String, DomElement> actions = generateActions(importPaths);
+        final Optional<Pair<String, DomElement>> actionPair = generateAction(importPaths, scatterPlotMasterBuilder.getAction());
 
         final String entityMasterStr = ResourceLoader.getText("ua/com/fielden/platform/web/master/scatter-plot/tg-scatter-plot-master-template.js")
                 .replace(IMPORTS, createImports(importPaths))
                 .replace(ENTITY_TYPE, flattenedNameOf(scatterPlotMasterBuilder.getEntityType()))
-                .replace("<!--@tg-entity-master-content-->", actions.getValue().toString())
-                .replace("//generatedPrimaryActions", actions.getKey())
+                .replace("<!--@tg-entity-master-content-->", actionPair.map(a -> a.getValue().toString()).orElse(""))
+                .replace("//generatedPrimaryActions", actionPair.map(a -> a.getKey().toString()).orElse(""))
                 .replace("//@ready-callback", readyCallback(scatterPlotMasterBuilder))
                 .replace("@prefDim", "null")
                 .replace("@noUiValue", "false")
                 .replace("@saveOnActivationValue", String.valueOf(scatterPlotMasterBuilder.shouldSaveOnActivation()));
 
-        renderable = new IRenderable() {
+        this.renderable = new IRenderable() {
             @Override
             public DomElement render() {
                 return new InnerTextElement(entityMasterStr);
@@ -68,19 +74,13 @@ public class ScatterPlotMaster<T extends AbstractEntity<?>> implements IMaster<T
         };
     }
 
-    private Pair<String, DomElement> generateActions(final LinkedHashSet<String> importPaths) {
-        final DomElement container = new DomContainer();
-        final List<String> primaryActionObjects = new ArrayList<>();
-        for (int actionIdx = 0; actionIdx < actions.size(); actionIdx++) {
-            final EntityActionConfig config = actions.get(actionIdx);
-            if (config != null) {
-                final FunctionalActionElement el = FunctionalActionElement.newEntityActionForMaster(config, actionIdx);
-                importPaths.add(el.importPath());
-                container.add(el.render().clazz("chart-action").attr("hidden", true).attr("action-index", actionIdx));
-                primaryActionObjects.add(el.createActionObject());
-            }
+    private Optional<Pair<String, DomElement>> generateAction(final LinkedHashSet<String> importPaths, final EntityActionConfig action) {
+        if (action != null) {
+            final FunctionalActionElement el = FunctionalActionElement.newEntityActionForMaster(action, 0);
+            importPaths.add(el.importPath());
+            return of(new Pair<>(el.createActionObject(), el.render().clazz("chart-action").attr("hidden", true)));
         }
-        return new Pair<>(StringUtils.join(primaryActionObjects, ",\n"), container);
+        return empty();
     }
 
     private String readyCallback(final ScatterPlotMasterBuilder<T> chartBuilder) {
@@ -118,17 +118,17 @@ public class ScatterPlotMaster<T extends AbstractEntity<?>> implements IMaster<T
     }
 
     private String getYAxisTitle(final ScatterPlotMasterBuilder<T> chartBuilder) {
-        return Optional.ofNullable(chartBuilder.getYAxisTitle())
+        return ofNullable(chartBuilder.getYAxisTitle())
                 .orElse(getTitleAndDesc(chartBuilder.getCategoryPropertyName(), chartBuilder.getChartEntityType()).getKey());
     }
 
     private String getXAxisTitle(final ScatterPlotMasterBuilder<T> chartBuilder) {
-        return Optional.ofNullable(chartBuilder.getXAxisTitle())
+        return ofNullable(chartBuilder.getXAxisTitle())
                 .orElse(getTitleAndDesc(chartBuilder.getValuePropertyName(), chartBuilder.getChartEntityType()).getKey());
     }
 
     private String getChartTitle(final ScatterPlotMasterBuilder<T> chartBuilder) {
-        return Optional.ofNullable(chartBuilder.getTitle())
+        return ofNullable(chartBuilder.getTitle())
                 .orElse(getEntityTitleAndDesc(chartBuilder.getChartEntityType()).getKey());
     }
 
@@ -166,15 +166,15 @@ public class ScatterPlotMaster<T extends AbstractEntity<?>> implements IMaster<T
 
     @Override
     public EntityActionConfig actionConfig(final FunctionalActionKind actionKind, final int actionNumber) {
-        if (PRIMARY_RESULT_SET.equals(actionKind)) {
-            return this.actions.get(actionNumber);
+        if (PRIMARY_RESULT_SET.equals(actionKind) && actionNumber == 0) {
+            return this.action;
         }
         throw new UnsupportedOperationException("Getting of action configuration is not supported.");
     }
 
     @Override
     public Optional<Class<? extends IValueMatcherWithContext<T, ?>>> matcherTypeFor(final String propName) {
-        return Optional.empty();
+        return empty();
     }
 
 }
