+++ platform-web-ui/src/main/web/ua/com/fielden/platform/web/master/scatter-plot/tg-scatter-plot-master-behavior.js
@@ -1,5 +1,6 @@
-import { TgMasterWithChartBehavior} from '/resources/master/tg-master-with-chart-behavior.js';
+import { TgEntityMasterTemplateBehavior} from '/resources/master/tg-entity-master-template-behavior.js'
 import { TgAppConfig } from '/app/tg-app-config.js';
+import {_momentTz} from '/resources/reflection/tg-date-utils.js';
 
 const getTooltipValueForEntity = function (entity) {
     const keyValue = entity.toString();
@@ -54,11 +55,31 @@ const TgScatterPlotMasterBehaviorImpl = {
     },
 
     ready: function () {
+
+        this.actions = [...this._masterDom().querySelectorAll(".chart-action")];
+
         this._appConfig = new TgAppConfig();
         this._click = this._click.bind(this);
         this._tooltip = this._tooltip.bind(this);
         this._getCategoryRange = this._getCategoryRange.bind(this);
         this._getValueRange = this._getValueRange.bind(this);
+    },
+
+    _datePropAccessor: function (propertyName, dateType) {
+        return function (entity, value) {
+            if (!value) {
+                const splitedType = dateType.split(':');
+                return _momentTz(entity.get(propertyName), splitedType[1] || null, splitedType[2] || null).toDate();
+            }
+        }
+    },
+
+    _moneyPropAccessor: function (propertyName) {
+        return function (entity, value) {
+            if (!value) {
+                return entity.get(propertyName).amount;
+            }
+        }.bind(this);
     },
 
     _getCategoryRange: function () {
@@ -112,6 +133,6 @@ const TgScatterPlotMasterBehaviorImpl = {
 };
 
 export const TgScatterPlotMasterBehavior = [
-    TgMasterWithChartBehavior,
+    TgEntityMasterTemplateBehavior,
     TgScatterPlotMasterBehaviorImpl
 ];
